Prepared deep koopman and feedforward code for experiments. Both versions print epoch number, runtime, training loss, validation loss, and comparison loss after 1, 5, and 15 time steps.
Changed deep koopman code to print every 10 epochs (it was previously set to print every 20 batches).
For both codes, the random seed is now set to 1, and both codes are set to run for 1 hour.

I also lowered the learning rate from 0.001 to 0.0005 for the feedforward code because the training loss was oscillating.