Updated DeepDMD.py to work on Lorenz system with rho=1 and 100 observables.
Made number of neurons per hidden layer a parameter that can be chhanged at the beginning of the file, called 'width'.
Made the string detailing the experiment number, called 'experiment_tag', another parameter that can be changed at the beginning of the file.
Trying network architecture 3 400 400 400 100 with training time 2 hours.